Add Animal class and its attributes.
This is part of the refactoring to our tP direction, where we want to
store animal details in the AddressBook rather than Person.
As such, the usage of Person will progressively be replaced with
completed implementation of Animal.

NOTE: The attributes of Animal are mostly incomplete as of this commit.